Move cr8tor core modules to airlock folder
