+++ src/cr8tor/airlock/api_client.py
@@ -4,7 +4,7 @@ from pydantic import BaseModel
 from typing import Optional, Union, Literal, Any, Dict
 from dotenv import load_dotenv, find_dotenv
 import json
-from cr8tor.core.schema import (
+from cr8tor.airlock.schema import (
     DataContractPublishRequest,
     DataContractValidateRequest,
     DataContractTransferRequest,
